Removed even more duplicated code
