+++ src/index.ts
@@ -1,10 +1,10 @@
 import type { PublicModel } from '@/src/types/model';
 import type { Query, Statement } from '@/src/types/query';
 import { compileQueryInput } from '@/src/utils';
-import { transformMetaQuery } from '@/src/utils/meta';
 import {
   addDefaultModelFields,
   addDefaultModelPresets,
+  addModelStatements,
   addSystemModels,
 } from '@/src/utils/model';
 
@@ -38,7 +38,7 @@ export const compileQueries = (
   for (const query of queries) {
     const statementValues = options?.inlineParams ? null : [];
 
-    const transformedQuery = transformMetaQuery(
+    const transformedQuery = addModelStatements(
       modelListWithPresets,
       dependencyStatements,
       statementValues,
